requirements.txt
numpy==1.24.4
Pillow==11.1.0
pyassimp==5.2.5
PyOpenGL==3.1.9
PyOpenGL==3.1.9
PyQt5==5.15.11
PyQt5_sip==12.15.0
